requirements.txt
pandas
numpy
matplotlib
streamlit
psycopg2-binary
reportlab
